+++ packages/docs/src/stories/Chart.stories.tsx
@@ -1,0 +1,9 @@
+import { Chart, ChartProps } from '@horus-finance/react';
+import { Meta, StoryObj } from '@storybook/react';
+
+export default {
+	title: 'Data display/Chart',
+	component: Chart,
+} as Meta<ChartProps>;
+
+export const Primary: StoryObj<ChartProps> = {};