+++ packages/docs/src/stories/Dropdown.stories.tsx
@@ -1,0 +1,28 @@
+// import {
+// 	Dropdown,
+// 	DropdownContent,
+// 	DropdownItem,
+// 	DropdownProps,
+// 	DropdownTrigger,
+// } from '@horus-finance/react';
+// import { StoryObj, Meta } from '@storybook/react';
+// import React from 'react';
+// import { ArrowsDownUp, CaretDown, ChartPieSlice, Gear } from 'phosphor-react';
+
+// export default {
+// 	title: 'Form/Dropdown',
+// 	component: Dropdown,
+// 	args: {
+// 		children: (
+// 			<>
+// 				<DropdownTrigger prefix="Customer: ">Jhon Doe</DropdownTrigger>
+// 				<DropdownContent width="203px">
+// 					<DropdownItem>Jhon Doe</DropdownItem>
+// 					<DropdownItem>Doe Jhon</DropdownItem>
+// 				</DropdownContent>
+// 			</>
+// 		),
+// 	},
+// } as Meta<DropdownProps>;
+
+// export const Primary: StoryObj<DropdownProps> = {};