+++ packages/react/src/utils/CurrencyConvert.ts
@@ -1,0 +1,7 @@
+export function currencyConvert(amount: number) {
+	const convertedAmount = amount.toLocaleString('pt-br', {
+		style: 'currency',
+		currency: 'BRL',
+	});
+	return convertedAmount;
+}