+++ packages/tokens/src/colors.ts
@@ -3,11 +3,15 @@ export const colors = {
 	black: '#000',
 
 	primary: '#1573FE',
+	background: '#15223C',
 
 	gray100: '#FAFAFA',
+	gray150: '#FEFEFE',
 	gray200: '#F5F5F5',
+	gray250: '#F5F6FA',
 	gray300: '#EEEEEE',
 	gray400: '#D9D9D9',
+	gray450: '#AFAFAF',
 	gray500: '#9E9E9E',
 	gray600: '#757575',
 	gray700: '#616161',
@@ -16,7 +20,10 @@ export const colors = {
 
 	blue100: '#E3F6FE',
 	blue200: '#24BEFF',
-	blue300: '#00B3FF', 
-	blue400: '#009CDE', 
+	blue300: '#00B3FF',
+	blue400: '#009CDE',
+
+	red500: '#E74A51',
+
+	green500: '#2DD683',
 };
-