+++ packages/ts-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @horus-finance/ts-config
 
+## 4.0.0
+
+### Major Changes
+
+- Finished all components first version
+
 ## 3.0.0
 
 ### Major Changes